+++ libs/core/src/components/fab/fab.ts
@@ -22,7 +22,6 @@ export class GdsFab extends GdsButton {
    */
   @styleExpressionProperty({
     valueTemplate: (v) => v,
-    selector: '[part="_button"]',
   })
   position?: string